remove fix from one spot to trigger warning again

+++ source/mqtt_request_response.c
@@ -604,7 +604,7 @@ static int s_compute_request_response_storage_properties(
 
     // Step 3 - Go through all the subscription topic filters, response paths, and options fields and add up
     // the lengths of all the string and binary data fields.
-    for (uint32_t i = 0; i < subscription_filter_count; ++i) {
+    for (size_t i = 0; i < subscription_filter_count; ++i) {
         napi_value array_element;
         AWS_NAPI_CALL(env, napi_get_element(env, node_subscription_topic_filters, i, &array_element), {
             AWS_LOGF_ERROR(
